Add tags option
The new option is a space-separated list of paths to search for tags
files. This is useful in cases where the tags file is kept out-of-tree,
for instance in a git repository's .git folder. In such a case, the
appropriate value for this setting looks like:

    set tags "tags .git/tags"

or, if you don't also want to search for tags in the standard location,
simply:

    set tags .git/tags

As this suggests, each entry in the list is searched in order.