Cleanup issues found with luacheck
Two functions were named get_path.  One to get the absolute path to
a file from ctags and one to get the path to the current file being
edited.  Change the names to abs_path and win_path to avoid confusion.

Change one variable name from path to abspath so it doesn't shadow
a function parameter.

Replace several unused variables with _ to indicate the fields
are unused.