Use quoted heredoc for safe shell usage
The use of echo -e "%s" could cause several problems:

- echo -e is non standard and could behave differently in different
shells, or even in the same shell compiled with different options.

- By placing the filenames in double quotes any filename with a $
or ` in it could cause an expansion, including code injection if the
filename had $(...) or `...` in it.

Instead, concat with literal newlines (instead of \n which has to be
interpreted), and pass to vis-menu via a quoted heredoc (<< 'EOF'
instead of << EOF) so nothing can expand and everything is passed
literally. Note that this still won't work with filenames that contain
newlines, but I don't think ctags will either and I know vis-menu
won't, so that's not worth worrying about.