Use vertical menu for tselect with extra info
Switch from vis-menu to vis-menu -l to show results vertically, one
file per line.  This is much easier to browse results.  Add the excmd
as well to give some context as to what the result is for a given file.
Break out the vis-menu command generation into a new function as
generating the command is no longer trivial.  Use an index to find the
match as we can't parse/compare the file and excmd after editing them.