Fix the inability to jump back after going to a symlinked file
When jumping to a tag that is present in a symlinked file, the path
being set in vis may not match the path stored in the tags file.
This will result in being unable to jump back because of the condition.
This commit uses the test command to resolve this; the flag '-ef'
isn't standard but commonly implemented in BSDs and Linux.
If you have OS X, Solaris, AIX... you may run into some difficulties
where the fallback behaviour will be handy.

Signed-off-by: Tai Chi Minh Ralph Eastwood <tcmreastwood@gmail.com>